# docs/offeredapis/swagger/pms-api.yaml
openapi: 3.0.1
info:
  title: A1 Policy management service
  description: <h2>General</h2><p>The O-RAN Non-RT RIC Policy Management Service provides
    a REST API for management of A1 policices. <br/>The main tasks of the service
    are:</p><ul><li>A1 Policy creation, modification and deletion.</li><li>Monitoring
    and maintaining consistency of the SMO view of A1 policies and the Near-RT RICs</li><li>Maintaining
    a view of supported Near-RT RIC policy types</li><li>Supervision of using services
    (R-APPs). When a service is unavailble, its policies are removed.</li></ul><h2>APIs
    provided by the service</h2><h3>A1 Policy Management</h3><p>This is an API for
    management of A1 Policies.</p><ul><li>A1 Policy retrieval, creation, modification
    and deletion.</li><li>Retrieval of supported A1 Policy types for a Near-RT RIC</li><li>Retrieval
    of status for existing A1 policies</li></ul><h3>Management of configuration</h3><p>API
    for updating and retrieval of the component configuration. Note that there other
    ways to maintain the configuration.</p><h3>Callbacks</h3><p>These are endpoints
    that are invoked by this service. The callbacks are registerred in this service
    at service registration.</p><h3>NearRT-RIC Repository</h3><p>This is an API that
    provides support for looking up a NearRT-RIC. Each A1 policy is targeted for one
    Near-RT RIC.</p><h3>Health Check</h3><p>API used for supervision of the PMS component.</p><h3>Service
...
  license:
...
servers:
...
paths:
...
  /services:
    get:
      tags:
      - A1 Policy Management V1.0
      summary: Returns service information
      operationId: getServices
      parameters:
      - name: name
        in: query
        description: The name of the service
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/service_status_v1'
        404:
          description: Service is not found
          content:
            '*/*':
              schema:
                type: string
    delete:
      tags:
      - A1 Policy Management V1.0
      summary: Delete a service
      operationId: deleteService
      parameters:
      - name: name
        in: query
        description: The name of the service
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        204:
          description: Service deleted
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/void'
        404:
          description: Service not found
          content:
            '*/*':
              schema:
                type: string
  /a1-policy/v2/rics/ric:
    get:
      tags:
      - NearRT-RIC Repository
      summary: Returns info for one Near-RT RIC
      description: Either a Near-RT RIC identity or a Mananged Element identity can
        be specified.<br>The intention with Mananged Element identity is the ID used
        in O1 for accessing the traffical element (such as the ID of CU).
      operationId: getRic_1
      parameters:
      - name: managed_element_id
        in: query
        description: The identity of a Managed Element. If given, the Near-RT RIC
          managing the ME is returned.
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: ric_id
        in: query
        description: The identity of a Near-RT RIC to get information for.
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: Near-RT RIC is found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ric_info_v2'
        404:
          description: Near-RT RIC is not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_information'
...
  /rics:
    get:
      tags:
      - A1 Policy Management V1.0
      summary: Query Near-RT RIC information
      operationId: getRics
      parameters:
      - name: policyType
        in: query
        description: The name of the policy type
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ric_info_v1'
        404:
          description: Policy type is not found
          content:
            '*/*':
              schema:
                type: string
...
  /a1-policy/v2/rics:
    get:
      tags:
      - NearRT-RIC Repository
      summary: Query Near-RT RIC information
      description: The call returns all Near-RT RICs that supports a given policy
        type identity
      operationId: getRics_1
      parameters:
      - name: policytype_id
        in: query
        description: The identity of a policy type. If given, all Near-RT RICs supporteing
          the policy type are returned
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ric_info_list_v2'
        404:
          description: Policy type is not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_information'
...
  /a1-policy/v2/services:
    get:
      tags:
      - Service Registry and Supervision
      summary: Returns service information
      description: Either information about a registered service with given identity
        or all registered services are returned.
      operationId: getServices_1
      parameters:
      - name: service_id
        in: query
        description: The identity of the service
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/service_list_v2'
        404:
          description: Service is not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_information'
    put:
      tags:
      - Service Registry and Supervision
      summary: Register a service
      description: Registering a service is needed to:<ul><li>Get callbacks.</li><li>Activate
        supervision of the service. If a service is inactive, its policies will be
        deleted.</li></ul>
      operationId: putService_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/service_registration_info_v2'
        required: true
      responses:
        200:
          description: Service updated
          content:
            '*/*':
              schema:
                type: object
        201:
          description: Service created
          content:
            '*/*':
              schema:
                type: object
        400:
          description: The ServiceRegistrationInfo is not accepted
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/error_information'
  /ric:
    get:
      tags:
      - A1 Policy Management V1.0
      summary: Returns the name of a RIC managing one Mananged Element
      operationId: getRic
      parameters:
      - name: managedElementId
        in: query
        description: The identity of the Managed Element
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: Near-RT RIC is found
          content:
            '*/*':
              schema:
                type: string
        404:
          description: Near-RT RIC is not found
          content:
            '*/*':
              schema:
                type: string
...
  /service:
    put:
      tags:
      - A1 Policy Management V1.0
      summary: Register a service
      operationId: putService
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/service_registration_info_v1'
        required: true
      responses:
        200:
          description: Service updated
          content:
            '*/*':
              schema:
                type: string
        201:
          description: Service created
          content:
            '*/*':
              schema:
                type: string
        400:
          description: The ServiceRegistrationInfo is not accepted
          content:
            '*/*':
              schema:
                type: string
  /a1-policy/v2/services/{service_id}:
    delete:
      tags:
      - Service Registry and Supervision
      summary: Unregister a service
      operationId: deleteService_1
      parameters:
      - name: service_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: Not used
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/void'
        204:
          description: Service unregistered
          content:
            '*/*':
              schema:
                type: object
        404:
          description: Service not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/error_information'
...
components:
  schemas:
    error_information:
      type: object
      properties:
        detail:
          type: string
          description: ' A human-readable explanation specific to this occurrence
            of the problem.'
          example: Policy type not found
        status:
          type: integer
          description: 'The HTTP status code generated by the origin server for this
            occurrence of the problem. '
          format: int32
          example: 503
      description: Problem as defined in https://tools.ietf.org/html/rfc7807
    void:
      type: object
      description: Void/empty
...
    service_registration_info_v1:
      type: object
      properties:
        keepAliveIntervalSeconds:
          type: integer
          description: keep alive interval for the service. This is a heartbeat supervision
            of the service, which in regular intevals must invoke a 'keepAlive' REST
            call. When a service does not invoke this call within the given time,
            it is considered unavailble. An unavailable service will be automatically
            deregistered and its policies will be deleted. Value 0 means no timeout
            supervision.
          format: int64
        callbackUrl:
          type: string
          description: callback for notifying of RIC synchronization
        serviceName:
          type: string
    ric_info_v2:
      type: object
      properties:
        ric_id:
          type: string
          description: identity of the Near-RT RIC
        managed_element_ids:
          type: array
          description: O1 identities for managed entities
          items:
            type: string
            description: O1 identities for managed entities
        state:
          type: string
          description: Represents the states for a Near-RT RIC
          enum:
          - UNAVAILABLE
          - AVAILABLE
          - SYNCHRONIZING
          - CONSISTENCY_CHECK
        policytype_ids:
          type: array
          description: supported policy types
          items:
            type: string
            description: supported policy types
      description: Information for a Near-RT RIC
    service_registration_info_v2:
      required:
      - service_id
      type: object
      properties:
        callback_url:
          type: string
          description: callback for notifying of Near-RT RIC state changes
        service_id:
          type: string
          description: identity of the service
        keep_alive_interval_seconds:
          type: integer
          description: keep alive interval for the service. This is a heartbeat supervision
            of the service, which in regular intevals must invoke a 'keepalive' REST
            call. When a service does not invoke this call within the given time,
            it is considered unavailble. An unavailable service will be automatically
            deregistered and its policies will be deleted. Value 0 means no timeout
            supervision.
          format: int64
      description: Information for one service
...
    ric_info_v1:
      type: object
      properties:
        managedElementIds:
          type: array
          description: O1 identities for managed entities
          items:
            type: string
            description: O1 identities for managed entities
        policyTypes:
          type: array
          description: supported policy types
          items:
            type: string
            description: supported policy types
        state:
          type: string
          description: state info
        ricName:
          type: string
          description: identity of the Near-RT RIC
    service_status_v1:
      type: object
      properties:
        keepAliveIntervalSeconds:
          type: integer
          description: policy keep alive timeout
          format: int64
        timeSinceLastActivitySeconds:
          type: integer
          description: time since last invocation by the service
          format: int64
        callbackUrl:
          type: string
          description: callback for notifying of RIC synchronization
        serviceName:
          type: string
          description: identity of the service
...
    service_status_v2:
      type: object
      properties:
        callback_url:
          type: string
          description: callback for notifying of RIC synchronization
        service_id:
          type: string
          description: identity of the service
        keep_alive_interval_seconds:
          type: integer
          description: policy keep alive timeout
          format: int64
        time_since_last_activity_seconds:
          type: integer
          description: time since last invocation by the service
          format: int64
      description: List of service information
    ric_info_list_v2:
      type: object
      properties:
        rics:
          type: array
          description: List of Near-RT RIC information
          items:
            $ref: '#/components/schemas/ric_info_v2'
      description: List of Near-RT RIC information
...
    service_list_v2:
      type: object
      properties:
        service_list:
          type: array
          description: List of service information
          items:
            $ref: '#/components/schemas/service_status_v2'
      description: List of service information
...